+++ include/ironbee/module_sym.h
@@ -118,11 +118,11 @@
 #ifdef __cplusplus
 /* C++ cannot do forward declarations for IB_MODULE_STRUCT. */
 #define IB_MODULE_DECLARE() \
-    ib_module_t DLL_PUBLIC *IB_MODULE_SYM(ib_engine_t* ib); \
+    const ib_module_t DLL_PUBLIC *IB_MODULE_SYM(ib_engine_t* ib); \
     extern ib_module_t IB_MODULE_STRUCT
 #else
 #define IB_MODULE_DECLARE() \
-    ib_module_t DLL_PUBLIC *IB_MODULE_SYM(ib_engine_t* ib); \
+    const ib_module_t DLL_PUBLIC *IB_MODULE_SYM(ib_engine_t* ib); \
     static ib_module_t IB_MODULE_STRUCT
 #endif
 
@@ -140,14 +140,14 @@
     ib_module_t IB_MODULE_STRUCT = { \
         __VA_ARGS__ \
     }; \
-    ib_module_t *IB_MODULE_SYM(ib_engine_t* ib) { return IB_MODULE_STRUCT_PTR; } \
+    const ib_module_t *IB_MODULE_SYM(ib_engine_t* ib) { return IB_MODULE_STRUCT_PTR; } \
     typedef int ib_require_semicolon_hack_
 #else
 #define IB_MODULE_INIT(...) \
     static ib_module_t IB_MODULE_STRUCT = { \
         __VA_ARGS__ \
     }; \
-    ib_module_t *IB_MODULE_SYM(ib_engine_t* ib) { return IB_MODULE_STRUCT_PTR; } \
+    const ib_module_t *IB_MODULE_SYM(ib_engine_t* ib) { return IB_MODULE_STRUCT_PTR; } \
     typedef int ib_require_semicolon_hack_
 #endif